requirements.txt
Django==3.0.9
djangorestframework==3.11.0
Pillow==7.2.0